added whitespace trimming and generalized gdb_to_postgres functions
