+++ sweri_utils/files.py
@@ -2,6 +2,7 @@ import os
 import arcpy
 import zipfile
 import requests
+import logging
 
 
 def export_file_by_type(fc_path, filetype, out_dir, out_name, tmp_path):
@@ -79,3 +80,39 @@ def extract_and_remove_zip_file(zip_filepath):
     with zipfile.ZipFile(zip_filepath, 'r') as zip_file:
         zip_file.extractall()
     os.remove(zip_filepath)
+
+def gdb_to_postgres(url, gdb_name, projection, fc_name, postgres_table_name, sde_file, schema):
+    # Downloads a gdb with a single feature class
+    # And uploads that featureclass to postgres
+    zip_file = f'{postgres_table_name}.zip'
+
+    #Download and extract gdb file
+    logging.info(f'Downloading {url}')
+    download_file_from_url(url, zip_file)
+
+    logging.info(f'Extracting {zip_file}')
+    extract_and_remove_zip_file(zip_file)
+ 
+    #Set Workspace to Downloaded GDB and set paths for feature class and reprojection
+    gdb_path = os.path.join(os.getcwd(),gdb_name)
+    feature_class = os.path.join(gdb_path,fc_name)
+    reprojected_fc = os.path.join(gdb_path, f'{postgres_table_name}')
+    postgres_table_location = os.path.join(sde_file, f'sweri.{schema}.{postgres_table_name}')
+
+    #Reproject layer
+    logging.info(f'reprojecting {feature_class}')
+    arcpy.Project_management(feature_class, reprojected_fc, projection)
+    logging.info('layer reprojected')
+
+    #Clear space in postgres for table
+    if(arcpy.Exists(postgres_table_location)):
+        arcpy.management.Delete(postgres_table_location)
+        logging.info(f'{postgres_table_name} has been deleted')
+
+    #Upload fc to postgres
+    arcpy.conversion.FeatureClassToGeodatabase(reprojected_fc, sde_file)
+    logging.info(f'{postgres_table_location} now in geodatabase')
+
+    #Remove gdb
+    arcpy.Delete_management(gdb_path)
+    logging.info(f'{gdb_path} deleted')
